% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apcDC.R
\name{apcDC}
\alias{apcDC}
\title{Adaptive Powered Correlation Prior with design covariates}
\usage{
apcDC(formula, design.formula, data, family = "gaussian", lambda = -1,
  log_lik = FALSE, iter = 10000, warmup = 1000, adapt = 5000,
  chains = 4, thin = 1, method = "parallel", cl = makeCluster(2),
  ...)
}
\arguments{
\item{formula}{the model formula}

\item{design.formula}{the formula for the design covariates}

\item{data}{a data frame}

\item{family}{one of "gaussian", "binomial", or "poisson".}

\item{lambda}{the power to use in the adaptive correlation prior. Default is -1, which gives the Zellner-Siow g prior. I suggest 
fitting multiple values of lambda and selecting which is best via LOO-IC or WAIC.}

\item{log_lik}{Should the log likelihood be monitored? The default is FALSE.}

\item{iter}{How many post-warmup samples? Defaults to 10000.}

\item{warmup}{How many warmup samples? Defaults to 1000.}

\item{adapt}{How many adaptation steps? Defaults to 2000.}

\item{chains}{How many chains? Defaults to 4.}

\item{thin}{Thinning interval. Defaults to 1.}

\item{method}{Defaults to "parallel". For an alternative parallel option, choose "rjparallel". Otherwise, "rjags" (single core run).}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods. Defaults to two cores.}

\item{...}{Other arguments to run.jags.}
}
\value{
A run.jags object.
}
\description{
This function adapts the adaptive powered correlation prior to the situation of estimating a single general(ized) linear regression
model. Furthermore this allows the inclusion of design variables separately from the set of numeric predictors.
\cr
\cr
Typically, in regression the cross-product XtX is inverted in the process of calculating the coefficients. In addition, 
the Zellner-Siow cauchy g-prior utilizes the inverse crossproduct is used as an empirical Bayesian method of determining the proper scale of the coefficient
priors by treating the inverse crossproduct as a covariance matrix, which is scaled by the parameter "g". \cr
\cr
The adaptive powered correlation prior simply extends this to allow using other powers besides -1. The power here will be referred to as "lambda".
Setting lambda to 0 results in a ridge-regression like prior. Setting lambda to a positive value adapts to collinearity by shrinking 
collinear variables towards a common value. Negative values of lambda pushes collinear variables further apart. Of course, setting lambda
to -1 is just the Zellner-Siow cauchy g-prior. This is designed to deal with collinearity in a more adaptive way than even ridge regression
by allowing the analyst to use model comparison techniques to choose an optimal value of lambda, and then using the best model for inference.
Note, however, that this prior is designed to deal with collinearity but not necessarily P > N scenarios. For that you may wish to take a look
at the \code{\link[Bayezilla]{glmBayes}} function, which does not utilize the crossproduct matrix in setting prior scales (rather they are
fully estimated in the model). 
\cr
}
\examples{
apcDC()

}
\references{
Krishna, A., Bondell, H. D., & Ghosh, S. K. (2009). Bayesian variable selection using an adaptive powered correlation prior. Journal of statistical planning and inference, 139(8), 2665–2674. doi:10.1016/j.jspi.2008.12.004
}
