% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apcLm.R
\name{apcLm}
\alias{apcLm}
\title{Adaptive Powered Correlation Prior for Gaussian likelihoods}
\usage{
apcLm(formula, data, lambda = -1, log_lik = FALSE, iter = 10000,
  warmup = 1000, adapt = 5000, chains = 4, thin = 1,
  method = "parallel", cl = makeCluster(2), ...)
}
\arguments{
\item{formula}{the model formula}

\item{data}{a data frame}

\item{lambda}{the power to use in the adaptive correlation prior. Default is -1, which gives the Zellner-Siow g prior. Setting
lambda to 0 results in a ridge-regression like prior. Setting lambda to a positive value adapts to collinearity by shrinking 
collinear variables towards a common value. Negative values of lambda pushes collinear variables further apart. I suggest 
fitting multiple values of lambda and selecting which is best via LOO-IC or WAIC.}

\item{log_lik}{Should the log likelihood be monitored? The default is FALSE.}

\item{iter}{How many post-warmup samples? Defaults to 10000.}

\item{warmup}{How many warmup samples? Defaults to 1000.}

\item{adapt}{How many adaptation steps? Defaults to 2000.}

\item{chains}{How many chains? Defaults to 4.}

\item{thin}{Thinning interval. Defaults to 1.}

\item{method}{Defaults to "parallel". For an alternative parallel option, choose "rjparallel". Otherwise, "rjags" (single core run).}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods. Defaults to two cores.}

\item{...}{Other arguments to run.jags.}

\item{family}{one of "gaussian", "binomial", or "poisson".}
}
\value{
A run.jags object.
}
\description{
Adaptive Powered Correlation Prior for Gaussian likelihoods
}
\examples{
apcLm()
}
