% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apcSpikeHS.R
\name{apcSpikeHS}
\alias{apcSpikeHS}
\title{Bernoulli-Normal Adaptive Powered Correlation Prior for Variable Selection}
\usage{
apcSpikeHS(formula, data, family = "gaussian", lambda = -1,
  log_lik = FALSE, iter = 8000, warmup = 2000, adapt = 1500,
  chains = 4, thin = 1, method = "rjparallel", cl = makeCluster(3),
  ...)
}
\arguments{
\item{formula}{the model formula}

\item{data}{a data frame}

\item{family}{one of "gaussian", "binomial", or "poisson".}

\item{lambda}{the power to use in the adaptive correlation prior. Default is -1, which gives the Zellner-Siow g prior. I suggest 
fitting multiple values of lambda and selecting which is best via LOO-IC or WAIC.}

\item{log_lik}{Should the log likelihood be monitored? The default is FALSE.}

\item{iter}{How many post-warmup samples? Defaults to 8000.}

\item{warmup}{How many warmup samples? Defaults to 2000.}

\item{adapt}{How many adaptation steps? Defaults to 1500.}

\item{chains}{How many chains? Defaults to 4.}

\item{thin}{Thinning interval. Defaults to 1.}

\item{method}{Defaults to "rjparallel". For an alternative parallel option, choose "parallel". Otherwise, "rjags" (single core run).}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods. Defaults to three cores.}

\item{...}{Other arguments to run.jags.}
}
\value{
A run.jags object.
}
\description{
IMPORTANT: I suggest not using any factor predictor variables, only numeric. In my experience the inclusion of categorical
predictors tends to lead to odd results in calculating the prior scale.

This function impements the adaptive powered correlation g-prior model described by Krishna, Bondell, and Ghosh (2009). Typically, in regression the cross-product XtX is inverted in the process of calculating the coefficients. In addition, 
the Zellner-Siow cauchy g-prior utilizes the inverse crossproduct is used as an empirical Bayesian method of determining the proper scale of the coefficient
priors by treating the inverse crossproduct as a covariance matrix, which is scaled by the parameter "g". 

The adaptive powered correlation prior simply extends this to allow using other powers besides -1. The power here will be referred to as "lambda".
Setting lambda to 0 results in a ridge-regression like prior. Setting lambda to a positive value adapts to collinearity by allowing
correlated predictors to enter and exit the model together. Negative values of lambda on the other hand favor including only one
of a set of correlated predictors. Of course, setting lambda to -1 is just the Zellner-Siow cauchy g-prior. This is designed to deal with collinearity in a more adaptive way than even ridge regression
by allowing the analyst to use model comparison techniques to choose an optimal value of lambda, and then using the best model for inference.
An analysts beliefs about which type of selection is preferred, or the goals of the particular analysis, can also inform the choice of lambda.

The difference between this and the apcSpike function is that there is a hyperprior on the overall inclusion probability prior.
This can be useful to encourage better sampling if the apcSpike function is having troubles, or you may wish
to use it if you do not wish to assume a beta prior on phi with fixed shape parameters. The structure of this hyperprior is as follows:

w ~ beta(1, 1) \cr
k_raw ~ uniform(0, 2) \cr
k <- k_raw + 2 \cr
a <- w * (K - 2) + 1 \cr
b <- (1-w) * (K - 2) + 1 \cr
phi ~ beta(a, b) \cr

for (i in 1:P){delta[i] ~ bernoulli(phi)} \cr

The specification of the k_raw as having a lower bound of zero, then adding 2 to k_raw only to subtract 2 from
k in the formula for the shape parameters a and b of the beta distribution might seem unnecessary, but coding it
this way prevents compilation errors.

Note, however, that this prior is designed to deal with collinearity but not necessarily P > N scenarios. For that you may wish to take a look
at the \code{\link[Bayezilla]{extLASSO}} or \code{\link[Bayezilla]{bayesEnet}} functions.
}
\examples{
apcSpikeHS()

}
\references{
Krishna, A., Bondell, H. D., & Ghosh, S. K. (2009). Bayesian variable selection using an adaptive powered correlation prior. Journal of statistical planning and inference, 139(8), 2665–2674. doi:10.1016/j.jspi.2008.12.004

Kuo, L., & Mallick, B. (1998). Variable Selection for Regression Models. Sankhyā: The Indian Journal of Statistics, Series B, 60(1), 65-81.
}
