% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesEnet.R
\name{bayesEnet}
\alias{bayesEnet}
\title{Bayesian Elastic Net for Gaussian Likelihood}
\usage{
bayesEnet(formula, data, family = "gaussian", log_lik = FALSE,
  iter = 10000, warmup = 1000, adapt = 2000, chains = 4,
  thin = 3, method = "rjags", cl = NULL, ...)
}
\arguments{
\item{formula}{the model formula}

\item{data}{a data frame.}

\item{log_lik}{Should the log likelihood be monitored? The default is FALSE.}

\item{iter}{How many post-warmup samples? Defaults to 10000.}

\item{warmup}{How many warmup samples? Defaults to 1000.}

\item{adapt}{How many adaptation steps? Defaults to 2000.}

\item{chains}{How many chains? Defaults to 4.}

\item{thin}{Thinning interval. Defaults to 3.}

\item{method}{Defaults to "rjags" (single core run). For parallel, choose "rjparallel" or "parallel".}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods.}

\item{...}{Other arguments to run.jags.}
}
\value{
A run.jags object
}
\description{
Bayesian Elastic Net for Gaussian Likelihood
}
\examples{
bayesEnet()

}
