% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_binomial.R
\name{binomTest}
\alias{binomTest}
\title{One or Two Sample Binomial Tests}
\usage{
binomTest(z, N, one.sample = TRUE, shape = c(0.5, 0.5), iter = 10000,
  warmup = 2500, adapt = 2500, chains = 4, thin = 3,
  method = "parallel", cl = makeCluster(2), summarise = FALSE, ...)
}
\arguments{
\item{z}{a numeric value giving the number of succesful trials. If two sample, a vector such as c(1, 3)}

\item{N}{a numeric value giving the total number of trials. If two sample, a vector such as c(5, 5)}

\item{one.sample}{Defaults to TRUE}

\item{shape}{the prior value for the shape parameters of the beta prior (defaults to c(0.50, 0.50))}

\item{iter}{the number of iterations. defaults to 10000.}

\item{warmup}{number of burnin samples. defaults to 2500.}

\item{adapt}{number of adaptation steps. defaults to 2500.}

\item{chains}{number of chains. defaults to 4.}

\item{thin}{the thinning interval. defaults to 3.}

\item{method}{Defaults to "parallel". For an alternative parallel option, choose "rjparallel" or. Otherwise, "rjags" (single core run).}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods. Defaults to two cores.}

\item{...}{other arguments to run.jags}
}
\value{
a runjags object
}
\description{
One or Two Sample Binomial Tests
}
\examples{
binomTest()

}
