% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blassoDC.R
\name{blassoDC}
\alias{blassoDC}
\title{Bayesian Lasso with unpenalized design covariates}
\usage{
blassoDC(formula, design.formula, data, log_lik = FALSE, iter = 10000,
  warmup = 1000, adapt = 2000, chains = 4, thin = 1,
  method = "parallel", cl = makeCluster(2), ...)
}
\arguments{
\item{formula}{the model formula.}

\item{design.formula}{formula for the design covariates.}

\item{data}{a data frame.}

\item{log_lik}{Should the log likelihood be monitored? The default is FALSE.}

\item{iter}{How many post-warmup samples? Defaults to 10000.}

\item{warmup}{How many warmup samples? Defaults to 1000.}

\item{adapt}{How many adaptation steps? Defaults to 2000.}

\item{chains}{How many chains? Defaults to 4.}

\item{thin}{Thinning interval. Defaults to 1.}

\item{method}{Defaults to "parallel". For an alternative parallel option, choose "rjparallel" or.
Otherwise, "rjags" (single core run).}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods. 
Defaults to two cores.}

\item{...}{Other arguments to run.jags.}
}
\value{
a runjags object
}
\description{
The Bayesian LASSO of Park & Casella (2008), but with the allowance for a set of covariates that are not penalized. 
For example, you may wish to include variables such as age and gender in all models so that 
the coefficients for the other variables are penalized while controlling for these. This
is a common need in research. \cr
\cr
Note only the Gaussian likelihood is provided because the Bayesian LASSO requires conditioning
on the error variance, which GLM-families do not have. \cr
}
\examples{
blassoDC()
}
\references{
Park, T., & Casella, G. (2008). The Bayesian Lasso. Journal of the American Statistical Association, 
103(482), 681-686. Retrieved from http://www.jstor.org/stable/27640090 \cr
}
