% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeGlm.R
\name{bridgeGlm}
\alias{bridgeGlm}
\title{Bridge Regression (Lq penalty)}
\usage{
bridgeGlm(formula, data, family = "gaussian", kappa = 0.5,
  log_lik = FALSE, iter = 4000, warmup = 1000, adapt = 5000,
  chains = 4, thin = 1, method = "parallel", cl = makeCluster(3),
  ...)
}
\arguments{
\item{formula}{the model formula}

\item{data}{a data frame.}

\item{family}{one of "gaussian", "binomial", or "poisson".}

\item{kappa}{the norm}

\item{log_lik}{Should the log likelihood be monitored? The default is FALSE.}

\item{iter}{How many post-warmup samples? Defaults to 10000.}

\item{warmup}{How many warmup samples? Defaults to 5000.}

\item{adapt}{How many adaptation steps? Defaults to 15000.}

\item{chains}{How many chains? Defaults to 4.}

\item{thin}{Thinning interval. Defaults to 3.}

\item{method}{Defaults to "parallel". For an alternative parallel option, choose "rjparallel" or. Otherwise, "rjags" (single core run).}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods. Defaults to two cores.}

\item{...}{Other arguments to run.jags.}
}
\value{
A run.jags object
}
\description{
This is the \Shape-adaptive shrinkage prior (SASP) 
model discussed by Mutshinda and Sillanp채채 (2011). 
\cr

Another model to consider is the  \code{\link[Bayezilla]{apcSpike}} (adaptive powered correlation prior) or 
Sillanp채채's other variable selection algorithms, the extended Bayesian Lasso and IAt models
provided in this package, respectively at \code{\link[Bayezilla]{extLASSO}} and \code{\link[Bayezilla]{IAt}} \cr
}
\examples{
SASP()

}
\references{
\cr
}
