% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCI.R
\name{compareCI}
\alias{compareCI}
\title{Check the sensitivity of credible intervals for the same data fit to different models to different priors}
\usage{
compareCI(list, model.names = "default", adj = 0.75)
}
\arguments{
\item{list}{a list of tidy summaries where the summaries are all of the same model with the same parameters. This
may require some preparation on your part.}

\item{model.names}{an optional character vector of model names.}

\item{adj}{the amount of space between intervals, deaults to .75}
}
\description{
Check the sensitivity of credible intervals for the same data fit to different models to different priors
}
\examples{
s1 = post_summary(model1, keeppars = c("Intercept", "beta"))
s2 = post_summary(model2, keeppars = c("Intercept", "beta"))
s3 = post_summary(model3, keeppars = c("Intercept", "beta"))
compareCI(list(s1, s2, s3))

}
