% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{confusion.matrix}
\alias{confusion.matrix}
\title{Binomial Confusion matrix for classification problems}
\usage{
confusion.matrix(predictions, actual)
}
\arguments{
\item{predictions}{the predicted classes.}

\item{actual}{the real outcome variable.}
}
\description{
An alternative to the caret package's confusion matrix function.
}
\examples{
confusion.matrix()

}
