% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corMat.R
\name{corMat}
\alias{corMat}
\title{Estimate a Correlation Matrix}
\usage{
corMat(x, iter = 10000, warmup = 2500, adapt = 2500, chains = 4,
  thin = 3, method = "parallel", cl = makeCluster(2), ...)
}
\arguments{
\item{x}{a data frame or matrix containing ONLY numeric variables to be correlated}

\item{iter}{the number of iterations. defaults to 10000.}

\item{warmup}{number of burnin samples. defaults to 2500.}

\item{adapt}{number of adaptation steps. defaults to 2500.}

\item{chains}{number of chains. defaults to 4.}

\item{thin}{the thinning interval. defaults to 3.}

\item{method}{Defaults to "parallel". For an alternative parallel option, choose "rjparallel" or. Otherwise, "rjags" (single core run).}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods. Defaults to two cores.}

\item{...}{other arguments to run.jags}
}
\value{
a runjags object
}
\description{
Model correlation matrices via the following model structure:

Tau ~ scaled wishart(s, 3) # Precision Matrix

Sigma <- inverse(Tau) # Covariance Matrix

sqrtSigma_1..i <- sqrt(Sigma[1..i, 1..i]) # Square Root of Diagonal

y ~ multivariate normal(emprical.means , Tau) # Likelihood Function

Rho_[i,i] <- Sigma[i,j] / sqrtSigma[i]*sqrtSigma[j] # Standardize the Covariance Matrix
}
\examples{
corTest(iris[,1:4])

}
