% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corTest.R
\name{corTest}
\alias{corTest}
\title{Bivariate Correlation Test}
\usage{
corTest(x, y = NULL, iter = 10000, warmup = 2500, adapt = 2500,
  chains = 4, thin = 3, method = "parallel", cl = makeCluster(2),
  ...)
}
\arguments{
\item{x}{a data frame or matrix containing two variables to be correlated, OR a single vector to be paired with y.}

\item{y}{if x is a vector, put the second variable to be correlated here.}

\item{iter}{the number of iterations. defaults to 10000.}

\item{warmup}{number of burnin samples. defaults to 2500.}

\item{adapt}{number of adaptation steps. defaults to 2500.}

\item{chains}{number of chains. defaults to 4.}

\item{thin}{the thinning interval. defaults to 3.}

\item{method}{Defaults to "parallel". For an alternative parallel option, choose "rjparallel" or. Otherwise, "rjags" (single core run).}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods. Defaults to two cores.}

\item{...}{other arguments to run.jags}
}
\value{
a runjags object
}
\description{
Bivariate Correlation Test
}
\examples{
corTest(iris$Sepal.Width, iris$Petal.Length)

}
