% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utilities.R
\name{cred_interval}
\alias{cred_interval}
\title{Calculate Equal Tailed Quantile Intervals or Highest Density Intervals}
\usage{
cred_interval(object, cred.level = 0.9, method = "QI")
}
\arguments{
\item{object}{vector of posterior samples}

\item{cred.level}{the confidence level. defaults to  .90}

\item{method}{type of credible interval. One of  "HDI", or default "QI".}
}
\description{
Calculate Equal Tailed Quantile Intervals or Highest Density Intervals
}
\examples{
cred_interval()
}
