% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dal}
\alias{dal}
\title{Asymmetric Laplace probability density function}
\usage{
dal(x, mu = 0, scale = 1, k = 0.5, log = FALSE)
}
\arguments{
\item{x}{vector of observations}

\item{mu}{vector of location parameter values}

\item{scale}{vector of scale parameter values}

\item{k}{vector of quantile locations}

\item{log}{if TRUE, probabilities p are given as log(p).}
}
\value{
vector
}
\description{
Asymmetric Laplace probability density function
}
\examples{
dal()
}
