% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{densPlot}
\alias{densPlot}
\title{Kernel Density Plot}
\usage{
densPlot(x, rug = TRUE, col = "#45a3ffCC", border = "#000f1e",
  rug.col = "#62a3e2", main.title = "Density Plot",
  kernel = "optcosine", n = 3072)
}
\arguments{
\item{x}{the numeric vector of interest}

\item{rug}{should a rug of the data be plotted underneath the density? Defaults to TRUE.}

\item{col}{The fill color of the density curve. Defaults to "#45a3ffCC"}

\item{border}{The border color of the density curve. Defaults to "#000f1e"}

\item{rug.col}{The color of the data rug. Defaults to "#62a3e2".}

\item{main.title}{Defaults to "Density Plot".}

\item{kernel}{one of the kernels from density(). Defaults to "optcosine"}

\item{n}{number of points to generate the density grid. Defaults to 3072.}
}
\value{
a plot
}
\description{
Kernel Density Plot
}
\examples{
densPlot(rnorm(1000))
}
