% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dpowexp}
\alias{dpowexp}
\title{Power Exponential probability density function}
\usage{
dpowexp(x, mu = 0, sigma = 1, kappa = 2, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{mu}{vector of location parameter values}

\item{sigma}{vector of scale parameter values}

\item{kappa}{vector of shape parameter values}

\item{log}{if TRUE, probabilities p are given as log(p).}
}
\value{
vector
}
\description{
Power Exponential probability density function
}
\examples{
dpowexp()
}
