% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dshash}
\alias{dshash}
\title{Sinh-Arcsinh (SHASH) probability density function}
\usage{
dshash(x, mu = 0, scale = 1, alpha = 0, nu = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{mu}{vector of location parameter values}

\item{scale}{vector of scale parameter values}

\item{alpha}{vector of skewness parameter}

\item{nu}{vector of kurtosis parameter}

\item{log}{if TRUE, probabilities p are given as log(p).}
}
\value{
vector
}
\description{
Sinh-Arcsinh (SHASH) probability density function
}
\examples{
dshash()
}
