% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dst}
\alias{dst}
\title{Student-T probability density function}
\usage{
dst(x, nu = 3, mu = 0, scale = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{nu}{vector of normality parameter values}

\item{mu}{vector of location parameter values}

\item{scale}{vector of scale parameter values}

\item{log}{if TRUE, probabilities p are given as log(p).}
}
\value{
vector
}
\description{
Student-T probability density function
}
\examples{
dst()
}
