% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidenceRatio.R
\name{evidenceRatio}
\alias{evidenceRatio}
\title{Obtain evidence ratios (pseudo-Bayes Factors) comparing two models}
\usage{
evidenceRatio(modelA, modelB, method = "LOO")
}
\arguments{
\item{modelA}{the first model}

\item{modelB}{the second model}

\item{method}{one of "LOO" (the default) or "WAIC"}
}
\value{
a data frame
}
\description{
This function is used to compare two models. The output can be interpreted
model compared to the other. Instead of using marginal likelihoods this function utilizes
either the Watanabe-Akaike Information Criterion or the Leave-One-Out Cross-Validation Information
Criterion. The default is to use LOO-IC. See the IC() function for more information on how
these are calculated. Models with a larger evidence ratio are favored over the alternative.
Posterior model probabilities are also provided for each model to aid interpretation.
}
\examples{
evidenceRatio()
}
