% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pred.R
\name{extract_pred}
\alias{extract_pred}
\title{Extract predicted y values (posterior predictions)}
\usage{
extract_pred(x, predname = "ySim")
}
\arguments{
\item{x}{the stanfit or runjags object containing posterior predictions}

\item{predname}{default name is "ySim".}
}
\description{
Extract predicted y values (posterior predictions)
}
\examples{
extract_post_pred()

}
