% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utilities.R
\name{extractCormat}
\alias{extractCormat}
\title{Extract Point Estimate of Correlation Matrix}
\usage{
extractCormat(fit, type = "Rho")
}
\arguments{
\item{fit}{the stanfit or runjags object.}

\item{type}{either "Rho" or "parRho"  which extracts columns labeled "Rho" or "parRho" respectively.}
}
\value{
a matrix
}
\description{
Extract Point Estimate of Correlation Matrix
}
\examples{
extractCormat(out)
}
