% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractPost}
\alias{extractPost}
\title{Extract a posterior to a data frame}
\usage{
extractPost(fit, keeppars = NULL, droppars = c("log_lik", "ySim",
  "yTest", "deviance", "omega", "lambda"))
}
\arguments{
\item{fit}{the model}

\item{keeppars}{parameters to keep. Defaults to NULL (implying anything is kept that's not in droppars)}

\item{droppars}{parameters to drop. Defaults to  c("log_lik", "ySim", "yTest", "deviance", "omega", "lambda")}
}
\value{
a data frame
}
\description{
This lets you extract the contents of a posterior to a data frame. You can
select parameters to keep and drop.
}
\examples{
extract(out)
}
