% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pred.R
\name{extractPred}
\alias{extractPred}
\title{Extract predicted y values (posterior predictions)}
\usage{
extractPred(x, predname = "ySim")
}
\arguments{
\item{x}{the stanfit or runjags object containing posterior predictions}

\item{predname}{default name is "ySim".}
}
\description{
Extract predicted y values (posterior predictions)
}
\examples{
extractPred()

}
