% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facetPost.R
\name{facetPost}
\alias{facetPost}
\title{Plot Multiple Posterior Distributions}
\usage{
facetPost(fit, pars = NULL, droppars = c("log_lik", "ySim", "yTest",
  "deviance"), col = "blue", nrow = 4, ncol = 2, method = "QI",
  showMedian = FALSE, cred.level = 0.9)
}
\arguments{
\item{fit}{a stanfit or runjags object. This can be used as an alternative to paramSampleVec,
but you must specify which parameter you would like to plot. If using this argument be sure to type
fit = "yourmodel" so that the function knows it is not intended to be a vector.}

\item{pars}{variables to keep}

\item{droppars}{variables to drop. Defaults to c("log_lik", "ySim", "yTest", "deviance")}

\item{col}{the color scheme. One of "blue", "green" "red", or "purple".}

\item{nrow}{number of rows in the layout}

\item{ncol}{number of columns in the layout}

\item{method}{Quantile Intervals "QI" (the default) or highest density intervals "HDI"}

\item{showMedian}{Should the median be used instead of the mean?
comparison value, the second being the lower limit of the ROPE and the third being the upper limit of the ROPE.}

\item{cred.level}{The credibility level. Defaults to 90\% (.90).}
}
\value{
A plot
}
\description{
Plot Multiple Posterior Distributions
}
\examples{
facetPost()

}
