% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamBayes.R
\name{gamBayes}
\alias{gamBayes}
\title{Bayesian GAMs}
\usage{
gamBayes(formula, data, family = "gaussian", log_lik = FALSE,
  iter = 4000, warmup = 1000, adapt = 1000, chains = 4, thin = 1,
  method = "rjparallel", cl = makeCluster(2), ...)
}
\arguments{
\item{formula}{the model formula}

\item{data}{a data frame}

\item{family}{one of "gaussian", "st" (Student-t with nu=3), "binomial", or "poisson".}

\item{log_lik}{Should the log likelihood be monitored? The default is FALSE.}

\item{iter}{How many post-warmup samples? Defaults to 4000.}

\item{warmup}{How many warmup samples? Defaults to 1000.}

\item{adapt}{How many adaptation steps? Defaults to 1000.}

\item{chains}{How many chains? Defaults to 4.}

\item{thin}{Thinning interval. Defaults to 1.}

\item{method}{Defaults to "rjparallel". Otherwise, "rjags" (single core run).}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods. Defaults to two cores.}

\item{...}{Other arguments to run.jags.}

\item{df}{degrees of freedom for prior.}

\item{s}{The desired prior scale. Defaults to 1. Is automatically squared within the model so
select a number here on the standard deviation scale.}
}
\value{
A run.jags object
}
\description{
This implements generalized additive models (GAMs) via the basis functions
available in the splines and splines2 packages. For example, you can use bs(), ns(),
cSpline(), iSpline(), mSpline(), or the base R function poly() to generate basis functions
for the variables intended for non-linear modeling. The basis functions are regularized
using a hierarchical prior where the top-level shrinkage parameter lambda is given a 
gamma prior, and the predictor specific shrinkage parameters are given a DuMouchel's
prior. Note that predictor specific means for each variable a basis function represents,
i.e., if variable x1 is represented by a basis function with 6 degrees of freedom then
all 6 components share a single shrinkage parameter, akin to the Group LASSO. This is standard 
procedure for GAMs.
}
\examples{
gamBayes()


}
