% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gammaPars}
\alias{gammaPars}
\title{Parameterize a Gamma Distribution by Mean and SD}
\usage{
gammaPars(mean, sd, plot = TRUE)
}
\arguments{
\item{mean}{mean}

\item{sd}{standard deviation}

\item{plot}{defaults to TRUE}
}
\value{
a plot and/or parameters
}
\description{
Parameterize a Gamma Distribution by Mean and SD
}
\examples{
gammaPars(3, 3)
}
