% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmFlat.R
\name{glmFlat}
\alias{glmFlat}
\title{Bayesian UN-regularized GLMs}
\usage{
glmFlat(formula, data, family = "gaussian", log_lik = FALSE,
  iter = 10000, warmup = 2500, adapt = 2500, chains = 4,
  thin = 2, method = "parallel", cl = makeCluster(2), ...)
}
\arguments{
\item{formula}{the model formula}

\item{data}{a data frame}

\item{family}{one of "gaussian", "binomial", or "poisson".}

\item{log_lik}{Should the log likelihood be monitored? The default is FALSE.}

\item{iter}{How many post-warmup samples? Defaults to 10000.}

\item{warmup}{How many warmup samples? Defaults to 2500.}

\item{adapt}{How many adaptation steps? Defaults to 2500.}

\item{chains}{How many chains? Defaults to 4.}

\item{thin}{Thinning interval. Defaults to 1.}

\item{method}{Defaults to "parallel". For an alternative parallel option, choose "rjparallel" or. Otherwise, "rjags" (single core run).}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods. Defaults to two cores.}

\item{...}{Other arguments to run.jags.}
}
\value{
A run.jags object
}
\description{
This model utilizes extremely vague uniform priors on all parameters. 
These are essentially, for practical purposes, improper flat priors. While JAGS 
does not support an improper prior over the whole real line, the coefficients 
have normal(0, 1e-200) priors (precision of 1e-200 is equivalent to a standard 
deviation of 1e100) and the precision has a gamma(1e-200, 1e-200) prior for
the Gaussian likelihood. No data that 
would ever be observed would be anywhere near this scale, so this can be used 
with scaled or unscaled data with equal validity, though I still recommend 
standardizing because MCMC samplers work better with data all on the same scale. 
The results from this should be equivalen to using lm() or glm()
in R.


Standard gaussian, binomial, and poisson likelihood functions are available.
}
\examples{
glmBayes()

}
\seealso{
\code{\link[Bayezilla]{apcGlm}}
}
