% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_extLASSO.R
\name{group_extLASSO}
\alias{group_extLASSO}
\title{Extended Group Bayesian LASSO}
\usage{
group_extLASSO(X, y, idx, family = "normal", eta_prior = "gamma",
  fixed_u = NA, log_lik = FALSE, iter = 10000, warmup = 1000,
  adapt = 2000, chains = 4, thin = 3, method = "rjags",
  cl = NULL, ...)
}
\arguments{
\item{X}{the model matrix. Construct this manually with model.matrix()[,-1]}

\item{y}{the outcome variable}

\item{idx}{the group labels. Should be of length = to ncol(model.matrix()[,-1]) with the group assignments
for each covariate. Please ensure that you start numbering with 1, and not 0.}

\item{family}{one of "gaussian", "binomial", or "poisson".}

\item{eta_prior}{one of "gamma" (default), "fixed_u", or "classic".}

\item{fixed_u}{if using "fixed_u" or "classic" this must be assigned a value. It is empty by
default to force you to choose a good value through model comparison.}

\item{log_lik}{Should the log likelihood be monitored? The default is FALSE.}

\item{iter}{How many post-warmup samples? Defaults to 10000.}

\item{warmup}{How many warmup samples? Defaults to 1000.}

\item{adapt}{How many adaptation steps? Defaults to 2000.}

\item{chains}{How many chains? Defaults to 4.}

\item{thin}{Thinning interval. Defaults to 3.}

\item{method}{Defaults to "rjags" (single core run). For parallel, choose "rjparallel" or "parallel".}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods.}

\item{...}{Other arguments to run.jags.}
}
\value{
A run.jags object
}
\description{
An adaptation of the extended Bayesian LASSO presented by Crispin M. Mutshinda and Mikko J. Sillanp채채 (2010)
for the purposes of group variable selection. The model specification is identical to the extended
Bayesian Lasso in all regards except the individual shrinkage hyperparameters & associated inclusion
indicators are replaced by group level parameters.

Three versions of the model are presented. The default is "gamma" which places a gamma(.01, .01) prior
on the top-level shrinkage hyperparameter and independent gamma(1.007137, 0.7000579) priors on the
individual group shrinkage hyperparameters. This prior has a median of exactly 1, making the prior
inclusion probability 50%. This leads to a denominator in the Bayes Factor for each coefficient of .50/.50 = 1,
so that the posterior odds are equivalent to the Bayes Factor. In other words the posterior is fully
data driven. This prior also can be a little better with convergence and sampling effeciency on certain data sets.

The second version is the "fixed_u" prior. While this retain the gamma(.01, .01) prior on the
top level shrinkage hyperparameter, the individaul shrinkage priors are of uniform(0, u) just as
in the original extended Bayesian LASSO. This requires you to choose an upper limit to the uniform
prior. Note that the prior inclusion probability is given by 1/u, so if you want a 50% inclusion
probability choose u=2. Common in Bayesian variable selection is to use a 20% probability if
dealing with a high dimensional problem, so for this choose u=5. If you have genuine prior
information you can and should use this to guide your choice. If you are unsure, use model comparison
to select which value of u to choose.

The third version of the model is the original specification in Mutshinda and Sillanp채채 (2010), albeit with
the group selection modification, labeled
"classic" in the options here. The only differing feature from the above is that the top level shrinkage
hyperparameter is given a uniform(0, 100) prior. The advice for the "fixed_u" version applies here as well.
}
\examples{
group_extLASSO
}
