% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_glm_spike.R
\name{group_glm_spike}
\alias{group_glm_spike}
\title{Bernoulli-Normal Mixture Group Selection for GLMs}
\usage{
group_glm_spike(formula, data, family = "gaussian", phi_prior = c(0.5,
  0.5), log_lik = FALSE, iter = 10000, warmup = 1000, adapt = 2000,
  chains = 4, thin = 3, method = "rjags", cl = NA, ...)
}
\arguments{
\item{family}{one of "gaussian", "binomial", or "poisson".}

\item{phi_prior}{The beta distribution parameters on the inclusion probabilities. Default is Jeffrey's prior, c(0.5, 0.5).}

\item{log_lik}{Should the log likelihood be monitored? The default is FALSE.}

\item{iter}{How many post-warmup samples? Defaults to 10000.}

\item{warmup}{How many warmup samples? Defaults to 1000.}

\item{adapt}{How many adaptation steps? Defaults to 2000.}

\item{chains}{How many chains? Defaults to 4.}

\item{thin}{Thinning interval. Defaults to 3.}

\item{method}{Defaults to "rjags" (single core run). For parallel, choose "rjparallel" or "parallel".}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods.}

\item{...}{Other arguments to run.jags.}

\item{X}{the model matrix. Construct this manually with model.matrix()[,-1]}

\item{y}{the outcome variable}

\item{idx}{the group labels. Should be of length = to ncol(model.matrix()[,-1]) with the group assignments
for each covariate. Please ensure that you start numbering with 1, and not 0.}
}
\value{
A run.jags object
}
\description{
This is the most basic type of Bayesian variable selection, however, it is
very intuitive to understand and often performs well**. This models the
regression coefficients as coming from either a null distribution represented
by a probability mass of 100% at zero, or as coming from a cauchy distribution parameterized
as a normal-gamma(.5,.5) mixture with the mixing parameter given the name "omega". This variant of the
Bernoulli-Normal mixture model models the selection of parameters as groups, akin to the group LASSO.

Note that you should center and scale your variables before using this function.
If you do not scale and center your numeric predictors, this will likely not perform well or
give reasonable results. The mixing hyperparameter omega assumes all covariates are on the same scale.


** This model works best on smaller to medium sized data sets. If you experience difficulty with
running times or obtaining a good effective sample size consider using the extended LASSO. Another tip that
may improve performance is using a beta(1, 1) or beta(1.5, 1.5) prior on phi. If all coefficients are set to
zero and there is truly good reason to believe this is not correct (ie, you aren't just hunting for "statistical
significance", are you?) a prior such as beta(4, 2) may help. If there is no sparsity, and you have genuine reason
to believe there should be try a beta(2,8) prior (if not, ask yourself why are you using variable selection?
If it is to deal with collinearity or other foibles for least squares, consider trying the glm_bayes function instead).
}
\examples{
group_glm_spike()

}
