% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_glm_spike.R
\name{groupSpike}
\alias{groupSpike}
\title{Bernoulli-Normal Mixture Group Selection for GLMs}
\usage{
groupSpike(formula, data, family = "gaussian", phi_prior = c(1, 4),
  log_lik = FALSE, iter = 10000, warmup = 1000, adapt = 2000,
  chains = 4, thin = 3, method = "parallel", cl = makeCluster(2),
  ...)
}
\arguments{
\item{family}{one of "gaussian", "binomial", or "poisson".}

\item{phi_prior}{Default is c(1, 4).}

\item{log_lik}{Should the log likelihood be monitored? The default is FALSE.}

\item{iter}{How many post-warmup samples? Defaults to 10000.}

\item{warmup}{How many warmup samples? Defaults to 1000.}

\item{adapt}{How many adaptation steps? Defaults to 2000.}

\item{chains}{How many chains? Defaults to 4. Max allowed is 4.}

\item{thin}{Thinning interval. Defaults to 3.}

\item{method}{Defaults to "parallel". For an alternative parallel option, choose "rjparallel". Otherwise, "rjags" (single core run).}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods. Defaults to two cores.}

\item{...}{Other arguments to run.jags.}

\item{X}{the model matrix. Construct this manually with model.matrix()[,-1]}

\item{y}{the outcome variable}

\item{idx}{the group labels. Should be of length = to ncol(model.matrix()[,-1]) with the group assignments
for each covariate. Please ensure that you start numbering with 1, and not 0.}
}
\value{
A run.jags object
}
\description{
IMPORTANT NOTICE: This model works best on smaller to medium sized data sets with a small number of variables (less than 20).
If you experience difficulty with running times or obtaining a good effective sample size consider using the extended LASSO.

IMPORTANT NOTICE: Center and scale your predictors before using this function.
If you do not scale and center your numeric predictors, this will likely not give reasonable results. 

This is the most basic type of Bayesian variable selection. This models the
regression coefficients as coming from either a null distribution represented
by a probability mass of 100% at zero (the "spike"), or as coming from a Normal( mu=0 , sigma = sqrt(2) ) distribution. This variant of the
Bernoulli-Normal mixture prior models the selection of parameters as groups, akin to the group LASSO.


Some Suggestions for Priors on phi (the overall inclusion probability):

beta(0.5, 0.5) Jeffrey's Prior (Truly uninformative)

beta(1.0, 1.0) Laplace's Uniform Prior

beta(1.0, 4.0) Weakly Informative, mean probability = 0.20 [This is the default prior used here]

beta(2.0, 2.0) Weakly Informative, mean probability = 0.50 

beta(4.0, 1.0) Weakly Informative, mean probability = 0.80

beta(2.0, 8.0) Moderately Informative, Regularizing, mean probability = 0.20 

beta(4.0, 4.0) Moderately Informative, Regularizing, mean probability = 0.50

beta(8.0, 2.0) Moderately Informative, Regularizing, mean probability = 0.80
}
\examples{
groupSpike()

}
