% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{imageIncl}
\alias{imageIncl}
\title{image the inclusions of variables from glm_spike over MCMC iterations}
\usage{
imageIncl(model, col_labs = NULL, colors = c("grey", "darkgreen"), ...)
}
\arguments{
\item{model}{the runjags object}

\item{col_labs}{column labels}

\item{colors}{defaults to grey for absence and darkgreen for inclusion, c("grey", "darkgreen")}

\item{...}{other arguments to image()}
}
\value{
an image
}
\description{
image the inclusions of variables from glm_spike over MCMC iterations
}
\examples{
imageIncl()

}
