% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{imageSigns}
\alias{imageSigns}
\title{Plot the signs of regression coefficients over MCMC iterations.}
\usage{
imageSigns(model, col_labs = NULL, colors = c("#99004C", "#1c1c1c",
  "#0065CC"), ...)
}
\arguments{
\item{model}{the runjags object}

\item{col_labs}{column labels}

\item{colors}{defaults to magenta, dark grey, and blue, c("#99004C", "#1c1c1c", "#0065CC")}

\item{...}{other arguments to image}
}
\value{
an image
}
\description{
View the signs of the regression coefficients over a series of MCMC runs. Accepts
any runjags object with the coefficients named "beta"
}
\examples{
imageSigns()

}
