% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{jaynes_psi}
\alias{jaynes_psi}
\title{Compute Jaynes' Psi between two distributions}
\usage{
jaynes_psi(a, b)
}
\arguments{
\item{a}{the first distribution}

\item{b}{the second distribution}
}
\description{
Returns the Jaynes' Psi divergence
}
\examples{
jaynes_psi()
}
