% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{jeffreys_dist}
\alias{jeffreys_dist}
\title{Compute Jeffrey's Distance between two distributions}
\usage{
jeffreys_dist(a, b, base = exp(1))
}
\arguments{
\item{a}{the first distribution}

\item{b}{the second distribution}

\item{base}{exponent base, defaults to exp(1)}
}
\description{
Returns the Jeffrey's Distance
}
\examples{
jeffreys_dist()
}
