% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utilities.R
\name{jointMode}
\alias{jointMode}
\title{Approximate the maximum joint posterior density estimate}
\usage{
jointMode(x, keeppars = NULL, droppars = c("ySim", "log_lik",
  "Deviance", "delta", "BIC"))

jointMode(x, keeppars = NULL, droppars = c("ySim", "log_lik",
  "Deviance", "delta", "BIC"))
}
\arguments{
\item{x}{the set of posterior samples to be summarized in an runjags or stanfit object.}

\item{keeppars}{the list of specific variables to keep. Defaults to NULL.}

\item{droppars}{list of parameters to exclude from the calculation of the density. The default is c("ySim", "log_lik", "lp__", "Deviance", "BIC", "delta"),
but this should be adjusted to make sure that derived and generated quantities such as predictions, deviance, indicator variables, and so forth --
anything not directly part of the posterior distribution of the model parameters -- are removed. Note, however, that the function will still
return all quantities in the model. The droppars argument only applies to the density estimation itself.}

\item{x}{the set of posterior samples to be summarized in an runjags or stanfit object.}

\item{keeppars}{the list of specific variables to keep. Defaults to NULL.}

\item{droppars}{list of parameters to exclude from the calculation of the density. The default is c("ySim", "log_lik", "lp__", "Deviance", "BIC", "delta"),
but this should be adjusted to make sure that derived and generated quantities such as predictions, deviance, indicator variables, and so forth --
anything not directly part of the posterior distribution of the model parameters -- are removed. Note, however, that the function will still
return all quantities in the model. The droppars argument only applies to the density estimation itself.}
}
\value{
a vector of modes

a vector of modes
}
\description{
This function calculates the density estimate for each column of a 
joint posterior distribution, then takes the logarithm of the estimate densities. The log-
densities are then summed to give the row-wise log-posterior probability function. The 
row with the highest log-posterior density is returned which corresponds to the monte carlo
sample with the highest joint probability, or MAP estimate. Note, however, this is an approximation
to the MAP in the sense that MAP estimation proper is done through optimization.

This function calculates the kernel density estimate for each column of a 
joint posterior distribution, then takes the logarithm of the estimate densities. The log-
densities are then summed to give the row-wise log-posterior probability function. The 
row with the highest log-posterior density is returned which corresponds to the monte carlo
sample with the highest joint probability, or MAP estimate. Note, however, this is an approximation
to the MAP in the sense that MAP estimation proper is done through optimization.
}
\examples{
jointMode(fit)
jointMode(fit)
}
