% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{kld}
\alias{kld}
\title{Compute the kullback leibler divergence}
\usage{
kld(a, b, base = exp(1))
}
\arguments{
\item{a}{the first distribution}

\item{b}{the second distribution}

\item{base}{exponent base, defaults to exp(1)}
}
\description{
Compute the kullback leibler divergence
}
\examples{
kld()
}
