% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpred.R
\name{linpred}
\alias{linpred}
\title{Generate predicted values from posterior samples for new observations}
\usage{
linpred(samples, formula, newdata, link = "identity")
}
\arguments{
\item{samples}{a matrix of MCMC samples containing the intercept and coefficients}

\item{formula}{the formula used for the original model}

\item{newdata}{a data frame of new data}

\item{link}{the link function used for the original model. For normal, student's t, laplace
and similar use "identity". For poisson glms use "log". For binomial models use either
"logitProb" if you want the predicted probabilities or "logitBin" if you want the classifications
split into the binary outcome, such that predicted probabilities < .50 are 0 and prob. > .50
are 1.}
}
\value{
a matrix
}
\description{
Note that this just generates the posterior for the predicted means / expected values.
for a set of new observations. This is not a posterior predictive distribution generator,
which draws samples from the outcome distribution. Hence, credible intervals for 
predictions derived in this function represent the credible intervals for the 
mean regression line, not the posterior predictive distribution which would have
considerably wider intervals.
}
\examples{
linpred(posterior, Sepal.Length ~ ., iris[testset,], "identity" )

}
