% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmSolve.R
\name{lmSolve}
\alias{lmSolve}
\title{Least Squares with Moore-Penrose Inverse}
\usage{
lmSolve(formula, data, tolerance = 0.001)
}
\arguments{
\item{formula}{formula}

\item{data}{data frame}

\item{tolerance}{tolerance}
}
\value{
a vector
}
\description{
A fast solver for simple least squares models. Returns coefficients only.
Solves with moore-penrose pseudoinversion which is very fast and tolerant
to non-positive definite model matrices.
}
\examples{
lmSolve()

}
