% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lssBayes.R
\name{lssBayes}
\alias{lssBayes}
\title{Regression Models for Location, Scale, and Shape}
\usage{
lssBayes(formula, sigma.formula = ~1, nu.formula = ~1, data,
  family = "gaussian", log_lik = FALSE, iter = 10000,
  warmup = 1000, adapt = 2000, chains = 4, thin = 3,
  method = "parallel", cl = makeCluster(2), ...)
}
\arguments{
\item{formula}{the model formula}

\item{sigma.formula}{the formula for the scale parameter. Defaults to ~ 1 (intercept only)}

\item{nu.formula}{the formula for the normality parameter, if using family = "student_t". Defaults to ~ 1 (intercept only)}

\item{data}{a data frame}

\item{family}{one of "gaussian", "laplace"}

\item{log_lik}{Should the log likelihood be monitored? The default is FALSE.}

\item{iter}{How many post-warmup samples? Defaults to 10000.}

\item{warmup}{How many warmup samples? Defaults to 1000.}

\item{adapt}{How many adaptation steps? Defaults to 2000.}

\item{chains}{How many chains? Defaults to 4.}

\item{thin}{Thinning interval. Defaults to 3.}

\item{method}{Defaults to "parallel". For an alternative parallel option, choose "rjparallel" or. Otherwise, "rjags" (single core run).}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods. Defaults to two cores.}

\item{...}{Other arguments to run.jags.}
}
\value{
A run.jags object
}
\description{
This is a very limited set models for estimating generalized linear models for location, scale, and shape.
Essentially what one does is model the scale parameter (see the examples for usage and syntax).
This is very useful if you have a heteroscedastic relationship between the linear predictor 
and the response variable. One can also model the scale of the response as a function of
a grouping variable, such as that shown in the examples where the Species factor variable
is used to model the standard deviation. Due to the limited set of distributions available
in JAGS, the support for the glmlss / gamlss family of models is extremely limited. However,
for more flexibility check out the bamlss package, which allows estimation of a wide class
of likelihood functions including skewed normal, power exponential (generalized normal), and
many more. This package only offers "gaussian" (the default), "student_t", and "laplace".
}
\details{
NOTE: Make sure to center and scale the predictor variables. All priors on coefficients
for all model parameters are set to a default precision of .0625, implying a standard 
deviation of 4 on each coefficient's prior. This is weakly informative when the predictors
are unit-scale, but cease to be weakly informative when the predictors are not unit scale.
}
\examples{
lssBayes(Petal.Width ~ . - Species, sigma.formula = ~ Species, data = scale(iris))
lssBayes(Petal.Width ~ . - Sepal.Length - Species, sigma.formula = ~ Species, nu.formula = ~ Sepal.Length, data = scale(iris), family = "student_t")

}
