% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utilities.R
\name{marginalModes}
\alias{marginalModes}
\title{Get the marginal mode estimate of a posterior distribution}
\usage{
marginalModes(x, keeppars = c("Intercept", "beta"),
  droppars = c("ySim", "log_lik", "lp__"))
}
\arguments{
\item{x}{the set of posterior samples to be summarized in an runjags or stanfit object.}

\item{keeppars}{the list of specific variables to keep. Defaults to c("Intercept", "beta")}

\item{droppars}{list of parameters to exclude}
}
\value{
a vector of modes
}
\description{
Get the marginal mode estimate of a posterior distribution
}
\examples{
marginalModes(fit)
}
