% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{nclass.dhist}
\alias{nclass.dhist}
\title{Density-based adaptive width histogram bins}
\usage{
nclass.dhist(x, b = 2, rx = range(x))
}
\arguments{
\item{b}{slope. See paper for details. Defaults to 2.}

\item{rx}{range of data, if not taken from data.}

\item{nbins}{number of bins. See Denby & Mallows (2009).}
}
\value{
A function that takes a single parameter, a numeric x specifying
  the data for which breaks are needed, and returns a vector of breaks.
}
\description{
This algorithm has some notable
advantages from a statistical point of view. Regions of high density have not only taller bins (as is usual) 
but more narrow bins as well. Regions of lower denisty have not only shorter, but wider bins. This makes the 
probability density much more 
immediately obvious, and captures interesting features of heavy tails and skew with greater efficacy. Algorithms
that yield The former oversmooths in regions of high density, and is poor at identifying sharp peaks and multimodality. 
By contrast, the latter variety oversmooths in regions of low density and can mask outliers and the 
heavy tails of more leptokurtotic distributions. Alternatively, a function can be supplied which will compute the
intended number of breaks or the actual breakpoints as a function of x. For more information, see Denby & Mallows (2009).
}
\examples{
nclass.dhist()

}
\references{
L. Denby and C. Mallows. Variations on the histogram. Journal
  of Computational and Graphical Statistics, 18 (1):21-31, 2009.
  URL \url{http://pubs.amstat.org/doi/abs/10.1198/jcgs.2009.0002.}
}
\seealso{
\url{http://pubs.research.avayalabs.com/pdfs/ALR-2007-003-paper.pdf}
}
\author{
Lorraine Denby
}
