% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_numvar}
\alias{plot_numvar}
\title{plot a tally of how many variables should be included}
\usage{
plot_numvar(model, binwidth = 0.5)
}
\arguments{
\item{model}{a runjags object with inclusion indicators labeled "delta"}

\item{binwidth}{defaults to .5}
}
\value{
a plot
}
\description{
plot a tally of how many variables should be included
}
\examples{
plot_numvar()

}
