% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pips}
\alias{plot_pips}
\title{plot inclusion probabilities from glm_spike models}
\usage{
plot_pips(model, col_labs = NULL)
}
\arguments{
\item{model}{the model}

\item{col_labs}{variable names}
}
\value{
a plot
}
\description{
lollipop plot with ggplot2 of variable inclusion probabilities for glm_spike models.
}
\examples{
plot_pips()

}
