% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{plot the coefficients as a barplot}
\usage{
plot_spectrum(samps, name = "beta", xlimits = c(0, length(samps)))
}
\arguments{
\item{samps}{a data frame or matrix of posterior samples}

\item{name}{name of the coefficients. Defaults to "beta"}

\item{xlimits}{defaults to 0 through the length of the samples.}
}
\value{
a plot
}
\description{
plot the coefficients as a barplot
}
\examples{
plot_spectrum()

}
