% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChains.R
\name{plotChains}
\alias{plotChains}
\title{MCMC Diagnostic Plots}
\usage{
plotChains(fit, pars = NULL, droppars = c("ySim", "log_lik"),
  col = "blues", nrow = 4, ncol = 2)
}
\arguments{
\item{fit}{the stanfit or runjags object}

\item{pars}{the names of the parameters to be plotted.}

\item{col}{the color scheme. One of "blue" (the default), "darkblue", "purples", "reds",
"greens", "mix" , or "mix2"}

\item{nrow}{number of rows in the layout}

\item{ncol}{number of columns in the layout}
}
\value{
a base R plot
}
\description{
Displays traceplots and cumulative means. Draws on Edward Tufte's minimalist aesthetic
ideas for the design.
}
\details{
Traceplots should look very
fuzzy and chaotic. If they look more ordered and correlated that indicates a problem
with the sampler or model specification.

Cumulative mean plots display how quickly the model converges onto the expected value.
The chains should all converge onto the same value or very close value very quickly.
If they do not, they indicates you either need to draw more samples, use a longer
adaptation and/or burnin/warmup for the MCMC chains, or perhaps reconsider the model
if those do not work.
}
\examples{
plotChains()
}
