% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCI.R
\name{plotCI}
\alias{plotCI}
\title{Plot Credible Intervals of a Bayesian Model}
\usage{
plotCI(fit, keeppars = NULL, droppars = c("ySim", "log_lik", "lp__"),
  estimate = "mean", cred.level = 0.9, method = "QI", ROPE = NULL,
  rope.color = "black", SEbars = FALSE,
  SEbar.color = "darkslategray3", point.color = "firebrick2",
  point.size = 2.75, point.shape = 18, bar.color = "skyblue4",
  bar.size = 0.8, H0 = 0, ...)
}
\arguments{
\item{fit}{the model fit}

\item{keeppars}{variables to keep}

\item{droppars}{variables to drop. defaults to c("ySim", "log_lik", "lp__")}

\item{cred.level}{Confidence level. Defaults to .90.}

\item{method}{credible interval methods. Defaults to QI (quantile intervals).}

\item{ROPE}{If you would like, ROPE limits added to the plot.}

\item{rope.color}{rope color}

\item{SEbars}{Defaults to FALSE. If TRUE, superimposes 68.2 percent intervals over wider ones.}

\item{SEbar.color}{standard error bar color}

\item{point.color}{point color}

\item{point.size}{point size}

\item{point.shape}{default to 18 (diamond)}

\item{bar.color}{bar color}

\item{bar.size}{bar size}

\item{H0}{the null hypothesis. defaults to 0.}

\item{...}{other arguments to pass to post_summary, which is called internally}
}
\description{
Plot Credible Intervals of a Bayesian Model
}
\examples{
plotCI()

}
