% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDist.R
\name{plotDist}
\alias{plotDist}
\title{plot a density function over a histogram}
\usage{
plotDist(x = NULL, dfunc, args = list(), bins = 20, n = 1500,
  type = "irregular", xname = "χ", xlab = xname, ylab = NULL,
  xlim = NULL, color = "#f45342CC", size = 2, ...)
}
\arguments{
\item{x}{the data}

\item{dfunc}{The name of a density function (ie, dnorm, dt, etc)}

\item{args}{list of the additional argument of the function, ie, list(mu = 0, sigma = 1)}

\item{bins}{defaults to 20}

\item{n}{the number of x values to be evaluated in the superimposed density curve. Defaults to 1500.}

\item{type}{defaults to "irregular" using the binning method described in plotPost. Otherwise, "equal".}

\item{xname}{character; argument of the function in func, the name of the x axis}

\item{xlab}{label of the x-axis}

\item{ylab}{label of the y-axis}

\item{xlim}{restrict the range of the function}

\item{color}{the color of the line. defaults to "red"}

\item{size}{the size of the line. defaults to 2}

\item{...}{Additional arguments to stat_function()}
}
\description{
plot a density function over a histogram
}
\examples{
plotDist(dnorm, from = -8 to = 8)
}
