% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPips}
\alias{plotPips}
\title{plot inclusion probabilities from SSVS models}
\usage{
plotPips(model, col_labs = NULL, font = "serif")
}
\arguments{
\item{model}{the model}

\item{col_labs}{variable names}

\item{font}{font for the axis labels. Defaults to "serif"}
}
\value{
a plot
}
\description{
lollipop plot with ggplot2 of variable inclusion probabilities for SSVS models.
}
\examples{
plotPips()

}
