% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPost.R
\name{plotPost}
\alias{plotPost}
\title{Histogram of Posterior Distribution}
\usage{
plotPost(paramSampleVec, fit = NULL, param = NULL, xlab = NULL,
  col = "blue", ROPE = NULL, cred.level = 0.9, method = "QI",
  CItextPlace = 0.7, breaks = "dhist", showMedian = FALSE, ...)
}
\arguments{
\item{paramSampleVec}{a vector containing the posterior distribution}

\item{fit}{a stanfit or runjags object. This can be used as an alternative to paramSampleVec,
but you must specify which parameter you would like to plot. If using this argument be sure to type
fit = "yourmodel" so that the function knows it is not intended to be a vector.}

\item{param}{the name of the parameter in the stanfit or runjags object you want to plot.}

\item{xlab}{if desired, a custom x-axis label. If supplying a fit object the label will be
the parameter name, but you can override that. If paramSampleVec is supplied without a custom
label here, the x-axis will display the Greek theta symbol by default.}

\item{col}{the color scheme. One of "blue", "green" "red", or "purple".}

\item{ROPE}{If you would like to display a ROPE, enter vector of three numbers, the first being the
comparison value, the second being the lower limit of the ROPE and the third being the upper limit of the ROPE.}

\item{cred.level}{The credibility level. Defaults to 90\% (.90).}

\item{method}{Quantile Intervals "QI" (the default) or highest density intervals "HDI"}

\item{CItextPlace}{Adjust the position of the credible interval text if needed.}

\item{breaks}{a number, function name, or function.}

\item{showMedian}{Should the median be used instead of the mean?}

\item{...}{other arguments to hist()}
}
\value{
a histogram
}
\description{
This is a heavy modification of John Kruchke's function of the same name in the BEST
package. This however makes many aesthetic changes and has some different features from the
original function.
\cr
}
\examples{
plotPost()
}
