% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_binomial.R
\name{poisTest}
\alias{poisTest}
\title{Poisson Test}
\usage{
poisTest(x, t = c(1, 1), r = 1, shra = c(1e-06, 1e-06),
  one.sample = TRUE, iter = 10000, warmup = 2500, adapt = 2500,
  chains = 4, thin = 3, method = "parallel", cl = makeCluster(2),
  summarise = FALSE, ...)
}
\arguments{
\item{x}{number of events. A vector of length one or two.}

\item{t}{time base for event count. A vector of length one or two.}

\item{r}{comparison value (only applies to one sample) on scale of lambda. Ie, on the scale of lambda_raw/t. Defaults to 1 and should typically stay 1 unless you have a specific hypothesis.}

\item{shra}{the shape and rate parameters of the gamma prior. Defaults to c(1e-6, 1e-6) which is non-informative and yields
a posterior mean typically equal to the maximum likelihood estimate.}

\item{one.sample}{Defaults to TRUE}

\item{iter}{the number of iterations. defaults to 10000.}

\item{warmup}{number of burnin samples. defaults to 2500.}

\item{adapt}{number of adaptation steps. defaults to 2500.}

\item{chains}{number of chains. defaults to 4.}

\item{thin}{the thinning interval. defaults to 3.}

\item{method}{Defaults to "parallel". For an alternative parallel option, choose "rjparallel" or. Otherwise, "rjags" (single core run).}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods. Defaults to two cores.}

\item{...}{other arguments to run.jags}
}
\value{
an rjags object
}
\description{
Poisson Test
}
\examples{
poisTest()

}
