% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utilities.R
\name{post_tibble}
\alias{post_tibble}
\title{Convert posterior summary to tibble}
\usage{
post_tibble(x, newnames = NULL, newcol = "term")
}
\description{
Convert posterior summary to tibble
}
\examples{
post_tibble()
}
\keyword{internal}
