% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postOdds.R
\name{postOdds}
\alias{postOdds}
\title{Get the posterior odds between a point estimator and a null hypothesis}
\usage{
postOdds(fit, param = NULL, H0 = 0, method = "mean")
}
\arguments{
\item{fit}{a runjags or stanfit object}

\item{param}{the name of the parameter to be tested}

\item{H0}{a single value or a vector of values for the null hypothesis. Defaults to zero but this is not appropriate
for a binomial test. Be sure to pick a reasonable null hypothesis.}

\item{method}{whether the mean (default), "median", or a kernel density estimate of the "mode" should be used for the hypothesis test}
}
\description{
Evaluates the density at the null hypothesis vs the density at the estimate, and calculates the
posterior odds as P(H0 | Data) / P(theta_hat | Data). This function is not an estimate of the Bayes
Factor. The Savage-Dickey Density Ratio Bayes Factors evaluate P(H0 | Data) / P(H0) 
and characterize the change in probability between the null hypothesis in the prior and posterior. 
Rather, this characterizes the relative posterior probabilities between a null hypothesis and a chosen
point estimate and is not reliant on the prior distribution. \cr
\cr
Smaller values indicate greater evidence against the null. A suggestion: Values equal to or larger than 0.33 
can be regarded as no evidence against the null, values smaller than 0.33 but larger than 0.2 as weak evidence
against the null, 0.2 as moderate evidence against the null, 0.1 as strong evidence against the null, 
and smaller than 0.1 as very strong evidence against the null. However these are NOT arbitrary cutoffs
as in null-hypothesis-significance testing with p-values, but rather, rough guidelines indexing
a continuous measure of evidential strength.
}
\examples{
postOdds()

}
