% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pow}
\alias{pow}
\title{Raise a number to a power}
\usage{
pow(x, lambda = NULL)
}
\arguments{
\item{x}{a number}

\item{lambda}{a power}
}
\value{
a number
}
\description{
Ever use the command "pow(x, 3)" only to find that R does not have a pow function? I do this all the time
out of habit because both JAGS and Stan use this, and matlab has the similar "power" function (but power does something else in R). 
This function exists to save frustration if you use the pow command to raise a number to a power by avoiding 
annoying error messages.
}
\examples{
pow(3, 4)
}
