% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{ppowexp}
\alias{ppowexp}
\title{Power Exponential cumulative probability function}
\usage{
ppowexp(q, mu = 0, sigma = 1, kappa = 2, lower.tail = TRUE,
  log = FALSE)
}
\arguments{
\item{q}{vector of quantiles}

\item{mu}{vector of location parameter values}

\item{sigma}{vector of scale parameter values}

\item{kappa}{vector of shape parameter values}

\item{lower.tail}{if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]}

\item{log}{if TRUE, probabilities p are given as log(p).}
}
\value{
vector
}
\description{
Power Exponential cumulative probability function
}
\examples{
ppowexp()
}
