% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{prec}
\alias{prec}
\title{Compute the precision of a vector}
\usage{
prec(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the precision
}
\description{
An R function to compute the precision of a vector, 1 / var(x)
}
\examples{
prec(x)
}
