% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmSolve.R
\name{pseudoinverse}
\alias{pseudoinverse}
\title{pseudoinverse}
\usage{
pseudoinverse(X, tol = 0.001)
}
\arguments{
\item{X}{a matrix}

\item{tol}{tolerance}
}
\value{
a matrix
}
\description{
Solve for the inverse of a matrix
}
\examples{
pseudoinverse()

}
