% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{pshash}
\alias{pshash}
\title{Sinh-Arcsinh (SHASH) cumulative probability function}
\usage{
pshash(q, mu = 0, scale = 1, alpha = 0, kappa = 1,
  lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{mu}{vector of location parameter values}

\item{scale}{vector of scale parameter values}

\item{alpha}{vector of skewness parameter}

\item{kappa}{vector of kurtosis parameter}

\item{lower.tail}{if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]}

\item{log.p}{if TRUE, probabilities p are given as log(p).}

\item{v}{vector of quantiles}
}
\value{
vector
}
\description{
Sinh-Arcsinh (SHASH) cumulative probability function
}
\examples{
pshash()
}
