% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{pst}
\alias{pst}
\title{Student-T cumulative probability function}
\usage{
pst(q, nu = 3, mu = 0, scale = 1, lower.tail = TRUE, log = FALSE)
}
\arguments{
\item{q}{vector of quantiles}

\item{mu}{vector of location parameter values}

\item{scale}{vector of scale parameter values}

\item{lower.tail}{if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]}

\item{log}{if TRUE, probabilities p are given as log(p).}

\item{kappa}{vector of normality parameter values}
}
\value{
vector
}
\description{
Student-T cumulative probability function
}
\examples{
pst()
}
