% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{qshash}
\alias{qshash}
\title{Sinh-Arcsinh (SHASH) quantile function}
\usage{
qshash(p, mu = 0, scale = 1, alpha = 0, nu = 1,
  lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{mu}{vector of location parameter values}

\item{scale}{vector of scale parameter values}

\item{alpha}{vector of skewness parameter}

\item{nu}{vector of kurtosis parameter}

\item{lower.tail}{if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]}

\item{log.p}{if TRUE, probabilities p are given as log(p).}
}
\value{
vector
}
\description{
Sinh-Arcsinh (SHASH) quantile function
}
\examples{
qshash()
}
