% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{qst}
\alias{qst}
\title{Student-T quantile function}
\usage{
qst(p, nu = 3, mu = 0, scale = 1, lower.tail = TRUE, log = FALSE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{nu}{vector of normality parameter values}

\item{mu}{vector of location parameter values}

\item{scale}{vector of scale parameter values}

\item{lower.tail}{if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]}

\item{log}{if TRUE, probabilities p are given as log(p).}
}
\value{
vector
}
\description{
Student-T quantile function
}
\examples{
qst()
}
