% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{ral}
\alias{ral}
\title{Asymmetric Laplace random number generator}
\usage{
ral(n, mu = 0, scale = 1, k = 0.5)
}
\arguments{
\item{n}{number of observations}

\item{mu}{vector of location parameter values}

\item{scale}{vector of scale parameter values}

\item{q}{vector of quantiles}
}
\value{
vector
}
\description{
Asymmetric Laplace random number generator
}
\examples{
ral()
}
