% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcorr.R
\name{rcorr}
\alias{rcorr}
\title{Generate Random Correlation Matrices}
\usage{
rcorr(p, eta = 1, df = NULL, method = "lkj")
}
\arguments{
\item{p}{The number of variabes to simulate}

\item{eta}{The regularization parameter. Larger values result in smaller correlations. Only applies to method = "lkj"}

\item{df}{The degrees of freedom. Only applies to method = "wishart"}

\item{One}{of "lkj" (the default) or "wishart"}
}
\value{
a correlation matrix
}
\description{
Generate Random Correlation Matrices
}
\examples{
rcorr(10, 2)

}
