% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ROPE}
\alias{ROPE}
\title{Conduct a ROPE test}
\usage{
ROPE(fit, ROPE = c(-0.1, 0.1), H0 = 0, plot = TRUE, param = NULL,
  line.color = "black", interval.color = "#8c00ffCC",
  density.color = "#2CBAFFCC", rope.color = "#FF0000CC",
  interval.method = "QI", cred.level = 0.9)
}
\arguments{
\item{fit}{a stanfit or runjags object.}

\item{ROPE}{a numeric vector of two numbers. Defaults to c(-.1, .1), assuming a Cohen's d... Adjust to what is reasonable!}

\item{plot}{plot the output. defaults to TRUE.}

\item{param}{the name of the parameter in the stanfit or runjags object you want to plot.}

\item{line.color}{color marking H0. "black" is default.}

\item{interval.color}{defaults to purple/violet "#8c00ffCC"}

\item{density.color}{defaults to blue "#2CBAFFCC"}

\item{rope.color}{defaults to red "#FF0000CC"}

\item{interval.method}{One of "QI" (the default), "HDI"}

\item{cred.level}{the confidence level for interval.method. Defaults to  .90}
}
\description{
Conduct a test of practical equivalence by defining a region of practical equivalence (ROPE) around the null hypothesis H0.
This is most appropriate when you wish to test whether an effect is larger in magnitude than a cutoff that defines an effect
size that is practically unimportant. Hence, this is a test for practical significance and requires a well defined idea of what
practical is.
}
\examples{
ROPE()

}
