% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rpowexp}
\alias{rpowexp}
\title{Power Exponential random number generator}
\usage{
rpowexp(n, mu = 0, sigma = 1, kappa = 2)
}
\arguments{
\item{n}{number of observations}

\item{mu}{vector of location parameter values}

\item{sigma}{vector of scale parameter values}

\item{kappa}{vector of shape parameter values}
}
\value{
vector
}
\description{
Power Exponential random number generator
}
\examples{
rpowexp()
}
