% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rshash}
\alias{rshash}
\title{Sinh-Arcsinh (SHASH) random number generator}
\usage{
rshash(n, mu = 0, scale = 1, alpha = 0, kappa = 1)
}
\arguments{
\item{n}{number of observations}

\item{mu}{vector of location parameter values}

\item{scale}{vector of scale parameter values}

\item{alpha}{vector of skewness parameter}

\item{kappa}{vector of kurtosis parameter}
}
\value{
vector
}
\description{
Sinh-Arcsinh (SHASH) random number generator
}
\examples{
rshash()
}
