% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rst}
\alias{rst}
\title{Student-T random number generator}
\usage{
rst(n, nu = 3, mu = 0, scale = 1)
}
\arguments{
\item{n}{number of observations}

\item{nu}{vector of normality parameter values}

\item{mu}{vector of location parameter values}

\item{scale}{vector of scale parameter values}
}
\value{
vector
}
\description{
Student-T random number generator
}
\examples{
rst()
}
