% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale}
\alias{scale}
\title{Scale a data frame or matrix}
\usage{
scale(data, scale.type = 1)
}
\arguments{
\item{data}{a data frame or vector}

\item{scale.type}{one of "sd" (default) for mean centered standard deviation scaled, "medmad" for
median centered median absolute deviation scaled, "medsd" for median centered standard
deviation scaled, "mean" for mean centered but unscaled, or "med" for median centered
but unscaled. Also accepts 1, 2, 3, 4, and 5 respectively as input types.}
}
\value{
A data frame
}
\description{
Scale a data frame or matrix
}
\details{
an improvement of the base R scale function. Unlike R's standard scale function, this
allows for factor columns or character columns to be present in the data, and will simply
leave those untouched without throwing an error.
}
\examples{
scale(data)

}
