% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatPlot.R
\name{scatMat}
\alias{scatMat}
\title{Visualize your data with a scatterplot matrix}
\usage{
scatMat(x, smooth = TRUE, digits = 2, method = "spearman",
  pch = 19, lm = FALSE, cor = TRUE, jitter = FALSE, amount = 2,
  hist.col = "#0054f9A6", points.col = "#35a8f59E",
  smooth.col = "#663399BF", breaks = "FD", cex.cor = 1,
  cex.num = 1.5, smoother = TRUE, font = "serif", ...)
}
\arguments{
\item{x}{The data frame}

\item{smooth}{Should the smoother line be enabled?}

\item{digits}{Number of significant digits. defaults to 2.}

\item{method}{method used for correlations; one of "pearson", "spearman" (the default), or "kendall"}

\item{pch}{The symbol used for plotting}

\item{lm}{TRUE or FALSE (Defaults to FALSE)}

\item{cor}{TRUE or FALSE (Defaults to TRUE)}

\item{jitter}{TRUE or FALSE (Defaults to FALSE)}

\item{amount}{Amount of jittering}

\item{hist.col}{The histogram color. Defaults to "#0054f9A6".}

\item{points.col}{The color of the data points. Defaults to "#35a8f59E".}

\item{smooth.col}{The color of the smooth line. Defaults to "#663399BF"}

\item{breaks}{Breaks in histogram. Defaults to "FD" method.}

\item{cex.cor}{Size of points}

\item{cex.num}{The size of the numbers in the upper right triangle of the scatterplot. Defaults to 1.5}

\item{smoother}{If TRUE, then smooth.scatter the data points – slow but pretty with lots of subjects}

\item{font}{font for the axis labels. Defaults to "serif"}

\item{...}{other arguments}
}
\value{
a plot
}
\description{
This is a function to help quickly visualize the relationships among your data.
Like the other pre-themed plotting functions in this package, this utilizes a
minimalist Tufte inspired aesthetic. When smooth is true it also automatically detects factor variables
or variables with less than 4 unique values to avoid trying to fit a smooth line
to nominal data, which is typically ugly at best and uninformative at worst.
}
\examples{
scatMat(iris)
}
