% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utilities.R
\name{sign_probs}
\alias{sign_probs}
\title{sign_probs}
\usage{
sign_probs(model, labs = NULL)
}
\arguments{
\item{model}{the model}

\item{labs}{variable labels}
}
\value{
a data frame
}
\description{
Summarize the sign probabilities of the coefficients in a runjags object
}
\examples{
sign_probs()
}
