% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikePlot.R
\name{spikePlot}
\alias{spikePlot}
\title{Plot a large number of variable values as vertical lines}
\usage{
spikePlot(paramSamples, main.title = "", lwd = 2.5)
}
\arguments{
\item{paramSamples}{the parameters as a vector of point values}

\item{main.title}{defaults to blank ""}

\item{lwd}{line width. defaults to 2.5}
}
\value{
a plot
}
\description{
Plot a large number of variable values as vertical lines
}
\examples{
spikePlot()
}
