% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{square}
\alias{square}
\title{Square a number}
\usage{
square(x)
}
\arguments{
\item{x}{a number}
}
\value{
a number
}
\description{
Ever use the command "square(x)" only to find that R does not have a "square" function, despite the fact
that it has a "sqrt" function? I do this all the time out of habit because I frequently use that when coding in Stan. 
This function exists to save me (or you) frustration by not incurring annoying error messages if using square(x) 
out of habit.
}
\examples{
square(2)
}
