% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2theme.R
\name{theme_min}
\alias{theme_min}
\title{Tufte-inspired ggplot2 theme}
\usage{
theme_min(base_size = 14, base_family = "serif",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\arguments{
\item{base_size}{Default font size = 14}

\item{base_family}{Font family. Defaults to "serif"}

\item{base_line_size}{defaults to base_size/22}

\item{base_rect_size}{defaults to base_size/22}
}
\description{
Tufte-inspired ggplot2 theme
}
\examples{
theme_set(theme_min())
}
