% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmSolve.R
\name{XtXinv}
\alias{XtXinv}
\title{Obtain the inverse crossproduct of a matrix}
\usage{
XtXinv(X, tol = 0.001)
}
\arguments{
\item{X}{the matrix}

\item{tol}{tolerance}
}
\value{
a matrix
}
\description{
Returns the pseudoinverse of the crossproduct. Use instead of solve(crossprod(X))
}
\examples{
XtXinv()

}
