% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corMat.R
\name{corMat}
\alias{corMat}
\title{Correlation Matrix Estimation (also returns partial correlations)}
\usage{
corMat(x, df = "default", iter = 4000, warmup = 2500, adapt = 2500,
  chains = 4, thin = 1, method = "parallel", cl = makeCluster(2),
  ...)
}
\arguments{
\item{x}{a data frame or matrix}

\item{df}{degrees of freedom for wishart prior. Defaults to ncol(X) + 1}

\item{iter}{the number of iterations. defaults to 4000.}

\item{warmup}{number of burnin samples. defaults to 2500.}

\item{adapt}{number of adaptation steps. defaults to 2500.}

\item{chains}{number of chains. defaults to 4.}

\item{thin}{the thinning interval. defaults to 3.}

\item{method}{Defaults to "parallel". For an alternative parallel option, choose "rjparallel" or. Otherwise, "rjags" (single core run).}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods. Defaults to two cores.}

\item{...}{other arguments to run.jags}
}
\value{
a runjags object
}
\description{
Just a simple multivariate norma-wishart conjugate model that returns the standardized inverse precision matrix (correlation matrix)
and standardized precision matrix (partial correlations).
}
\examples{
corMat(iris$Sepal.Width, iris$Petal.Length)

}
