% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{cred_interval}
\alias{cred_interval}
\title{Calculate Equal Tailed or Highest Density Interval}
\usage{
cred_interval(object, cred.level = 0.9, method = "QI")
}
\arguments{
\item{object}{vector of posterior samples}

\item{cred.level}{the confidence level. defaults to  .90}

\item{method}{type of credible interval. One of "QI" or default "HDI".}
}
\description{
Calculate Equal Tailed or Highest Density Interval
}
\examples{
cred_interval()
}
