% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utilities.R
\name{extractCormat}
\alias{extractCormat}
\title{Extract Point Estimate of Correlation Matrix}
\usage{
extractCormat(fit)
}
\arguments{
\item{fit}{the stanfit or runjags object. Must contain only variables with the name "Rho[,]".}
}
\value{
a matrix
}
\description{
Extract Point Estimate of Correlation Matrix
}
\examples{
extractCormat(out)
}
