% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupSpike.R
\name{groupSpike}
\alias{groupSpike}
\title{Bernoulli-Normal Mixture Group Selection for GLMs}
\usage{
groupSpike(X, y, family = "gaussian", phi_prior = c(1, 4),
  log_lik = FALSE, iter = 10000, warmup = 1000, adapt = 2000,
  chains = 4, thin = 1, method = "parallel", cl = makeCluster(2),
  ...)
}
\arguments{
\item{X}{the model matrix. Construct this manually with model.matrix()[,-1]}

\item{y}{the outcome variable}

\item{family}{one of "gaussian", "binomial", or "poisson".}

\item{log_lik}{Should the log likelihood be monitored? The default is FALSE.}

\item{iter}{How many post-warmup samples? Defaults to 10000.}

\item{warmup}{How many warmup samples? Defaults to 1000.}

\item{adapt}{How many adaptation steps? Defaults to 2000.}

\item{chains}{How many chains? Defaults to 4.}

\item{thin}{Thinning interval. Defaults to 1.}

\item{method}{Defaults to "parallel". For an alternative parallel option, choose "rjparallel". Otherwise, "rjags" (single core run).}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods. Defaults to two cores.}

\item{...}{Other arguments to run.jags.}

\item{idx}{the group labels. Should be of length = to ncol(model.matrix()[,-1]) with the group assignments
for each covariate. Please ensure that you start numbering with 1, and not 0.}
}
\value{
A run.jags object
}
\description{
IMPORTANT NOTICE: This model works best on smaller to medium sized data sets with a small number of variables (less than 20).
If you experience difficulty with running times or obtaining a good effective sample size consider using the extended group LASSO.

IMPORTANT NOTICE: Center and scale your predictors before using this function.
If you do not scale and center your numeric predictors, this will likely not give reasonable results. 

The model is very similar to the \code{\link[Bayezilla]{glmSpike}} model. Each group receives its own inclusion 
probability prior "phi", which is in turn controlled by an overall inclusion probability parameterized as a hierarchical 
structure similar to that in \code{\link[Bayezilla]{apcSpikeHS}} but simplified.
}
\examples{
groupSpike()

}
