% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negLASSO.R
\name{negLASSO}
\alias{negLASSO}
\title{Negative-exponential-gamma LASSO}
\usage{
negLASSO(formula, data, family = "gaussian", log_lik = FALSE,
  iter = 10000, warmup = 1000, adapt = 2000, chains = 4,
  thin = 1, method = "parallel", cl = makeCluster(3), ...)
}
\arguments{
\item{formula}{the model formula}

\item{data}{a data frame.}

\item{family}{one of "gaussian", "binomial", or "poisson".}

\item{log_lik}{Should the log likelihood be monitored? The default is FALSE.}

\item{iter}{How many post-warmup samples? Defaults to 10000.}

\item{warmup}{How many warmup samples? Defaults to 1000.}

\item{adapt}{How many adaptation steps? Defaults to 2000.}

\item{chains}{How many chains? Defaults to 4.}

\item{thin}{Thinning interval. Defaults to 1.}

\item{method}{Defaults to "parallel". For another parallel option, choose "rjparallel" or "rjags" for a single core run.}

\item{cl}{Use parallel::makeCluster(# clusters) to specify clusters for the parallel methods. Defaults to 3.}

\item{...}{Other arguments to run.jags.}
}
\value{
A run.jags object
}
\description{
This implements the normal-exponential-gamma "hyperlasso" of Griffin & Brown (2011).
}
\examples{
negLASSO()

}
\references{
Griffin, J. E. and Brown, P. J. (2011), Bayesian Hyper‐LASSOs With Non-Convex Penalization. Australian & New Zealand Journal of Statistics, 53: 423-442. doi:10.1111/j.1467-842X.2011.00641.x
}
