% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotNumvar}
\alias{plotNumvar}
\title{Plot distribution of num. of included vars. over iterations of an SSVS model}
\usage{
plotNumvar(model, binwidth = 0.5)
}
\arguments{
\item{model}{a runjags object with inclusion indicators labeled "delta"}

\item{binwidth}{defaults to .5}
}
\value{
a plot
}
\description{
Plot distribution of num. of included vars. over iterations of an SSVS model
}
\examples{
plotNumvar()

}
