% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utilities.R
\name{post_summary}
\alias{post_summary}
\title{Get a tidy summary of a bayesian model}
\usage{
post_summary(x, digits = 2, estimate.method = "both", ess = TRUE,
  cred.level = 0.9, cred.method = "QI", keeppars = NULL,
  droppars = c("ySim", "log_lik", "lp__"), knitr = FALSE,
  type = "markdown", ...)
}
\arguments{
\item{x}{the set of posterior samples to be summarized}

\item{digits}{number of digits to round estimates to. Defaults to 2.}

\item{estimate.method}{one of "mean" or "median", or "both". If "mean" or "median" the
estimate will be in column title "estimate". If "both" (the default) the mean will be the "estimate" column and the median will be in the "median" column.}

\item{ess}{set to TRUE (default) to include ess}

\item{cred.level}{the credibility level. defaults to 0.90}

\item{cred.method}{one of "HDI" (highest density intervals) or "QI" (equal tailed \ quantile intervals).}

\item{keeppars}{the list of specific variables to keep if passing an runjags object.}

\item{droppars}{list of parameters to exclude}

\item{knitr}{if you want to return the summary with knitr's kable function set to TRUE. Default is FALSE.}

\item{type}{output type for kable. Defaults to "markdown"}

\item{...}{other arguments to pass to knitr::kable}
}
\value{
a tibble or knitr::kable output
}
\description{
Get summary including QI or HDI credible intervals
}
\examples{
post_summary()

}
