% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatPlot.R
\name{scatPlot}
\alias{scatPlot}
\title{Bivariate Scatterplot}
\usage{
scatPlot(formula, data = NULL, xlab = "default", ylab = "default",
  col = "blues", smooth = FALSE, smoother = FALSE, regline = TRUE,
  font = "serif", inset = 0.05, legend.pos = "topright",
  box.lty = 1, box.cex = 0.75)
}
\arguments{
\item{formula}{a formula of the form y ~ x}

\item{data}{a data frame if y and x are not vectors in the global environment or subsets of a data frame}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{col}{color scheme. one of "blue" (the default), "red", "green", or "purple"}

\item{smooth}{should a smoother line be added. Defaults to FALSE.}

\item{smoother}{If FALSE the data are represented as points. If TRUE, a smooth density is used. The latter is useful for large sample sizes.}

\item{regline}{should a regression line be added? Defaults to TRUE, with both robust and ordinary least squares lines being plotted.}

\item{font}{Defaults to "serif"}

\item{inset}{the inset of the legend. Defaults to 0.05}

\item{legend.pos}{the legend position. Defaults to "topright"}

\item{box.lty}{the linetype of the legend box outline}

\item{box.cex}{the size of the legend}
}
\value{
a plot
}
\description{
Bivariate Scatterplot
}
\examples{
scatPlot(Sepal.Width ~ Petal.Length, iris, xlab = "Sepal Width", ylab = "Petal Length", col = "purple")
}
