% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatPlot.R
\name{scatPlot}
\alias{scatPlot}
\title{Scatterplots with Marginal Histograms}
\usage{
scatPlot(x, y, xlab = "x", ylab = "y", col = "blues", hist = TRUE,
  x.breaks = "dhist", y.breaks = "dhist", smooth = FALSE,
  regline = TRUE, font = "serif", inset = 0.05,
  legend.pos = "topright", box.lty = 1, box.cex = 0.95)
}
\arguments{
\item{x}{the x-axis variable or a formula}

\item{y}{the y-axis variable or a data frame if x is a formula}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{col}{color scheme. one of "blue" (the default), "red", "green", or "purple"}

\item{x.breaks}{breaks in x axis histogram. defaults to 15.}

\item{y.breaks}{breaks in y axis histogram. defaults to 15.}

\item{smooth}{should a smoother line be added. Defaults to FALSE.}

\item{regline}{should a regression line be added? Defaults to TRUE, with both robust and ordinary least squares lines being plotted.}

\item{font}{Defaults to "serif"}

\item{inset}{the inset of the legend. Defaults to 0.05}

\item{legend.pos}{the legend position. Defaults to "topright"}

\item{box.lty}{the linetype of the legend box outline}

\item{box.cex}{the size of the legend}
}
\value{
a plot
}
\description{
Scatterplots with Marginal Histograms
}
\examples{
scatterPlot(iris$Sepal.Width, iris$Petal.Length, xlab = "Sepal Width", ylab = "Petal Length", col = "purple")
}
