% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatPlot.R
\name{scatPlot}
\alias{scatPlot}
\title{Bivariate Scatterplot}
\usage{
scatPlot(formula, data = NULL, xlab = "x", ylab = "y",
  col = "blues", smooth = FALSE, regline = TRUE, ci = FALSE,
  conf.level = 0.9, font = "serif")
}
\arguments{
\item{formula}{a formula of the form y ~ x}

\item{data}{a data frame if y and x are not vectors in the global environment or subsets of a data frame}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{col}{color scheme. one of "blue" (the default), "red", "green", or "purple"}

\item{smooth}{should a smoother line be added. Defaults to FALSE.}

\item{regline}{should a regression line be added? Defaults to TRUE, with both robust and ordinary least squares lines being plotted.}

\item{ci}{should confidence intervals be displayed? Defaults to FALSE.}

\item{conf.level}{Default confidence level is 0.90. Set to 0.682 for standard error bars.}

\item{font}{Defaults to "serif"}
}
\value{
a plot
}
\description{
Bivariate Scatterplot
}
\examples{
scatPlot(iris$Sepal.Width ~ iris$Petal.Length, xlab = "Sepal Width", ylab = "Petal Length", col = "purple")
}
