% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikePlot.R
\name{spikePlot}
\alias{spikePlot}
\title{Plot a large number of variable values as vertical lines}
\usage{
spikePlot(x, estimate = "mean", keeppars = "beta",
  droppars = c("ySim", "log_lik", "lp__"), main.title = "parameters",
  lwd = 2.5, cex.axis = 0.85)
}
\arguments{
\item{x}{the model}

\item{estimate}{"mean" (the default) or "median"}

\item{keeppars}{The list of specific variables to keep if passing an runjags object. Defaults to "beta"}

\item{droppars}{list of parameters to exclude}

\item{main.title}{defaults to "parameters"}

\item{lwd}{line width. defaults to 2.5}

\item{cex.axis}{axis text size. defaults to .85.}
}
\value{
a plot
}
\description{
Plot a large number of variable values as vertical lines
}
\examples{
spikePlot()
}
